Day2/requirements.txt
aistudio-sdk==0.3.8
annotated-types==0.6.0
anyio==4.6.2.post1
astor==0.8.1
bce-python-sdk==0.9.52
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
decorator==5.1.1
easyocr==1.7.1
filelock==3.16.1
fsspec==2024.10.0
future==0.18.3
h11==0.14.0
hf-xet==1.1.7
httpcore==1.0.5
httpx==0.27.2
huggingface_hub==0.25.2
idna==3.10
imageio==2.36.0
imagesize==1.4.1
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
modelscope==1.28.1
mpmath==1.3.0
networkx==3.3
ninja==1.11.1
numpy==1.26.4
# opencv-contrib-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging==24.2
paddleocr==3.3.2
paddlepaddle==3.2.2
# paddlex==2.1.0
pandas==2.2.3
pillow==10.4.0
prettytable==3.11.0
protobuf==4.25.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pypdfium2==4.25.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytesseract==0.3.13
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
ruamel.yaml==0.18.6
flask==3.0.3
ruamel.yaml.clib==0.2.8
safetensors>=0.6.0
scikit-image==0.24.0
scipy==1.13.1
setuptools==75.1.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sympy==1.13.2
tifffile==2023.9.26
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.5
typer-slim==0.12.5
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
wcwidth==0.2.13
